requirements.txt
python-dotenv
selenium
webdriver-manager
requests
pygsheets
argparse